+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.0] - 2020-10-31
 ### Added
 - add pug to html tasc to gulp
+- add audio files
 
 ## [0.1.0] - 2020-10-31
 ### Added